% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelo.R
\name{Modelo}
\alias{Modelo}
\title{Modelo}
\usage{
Modelo(a, b, c, d, e, f, g)
}
\arguments{
\item{a}{es el albedo de la superficie de la tierra}

\item{b}{Radiacion solar estra-atmosferica en la superficie}

\item{c}{radiacion solar dispersa}

\item{d}{radación solar total}

\item{e}{angulo de inclinacion de la superficie horizontal}

\item{f}{angulo entre la direccion de la radiacion iniciente a la superficie y la normal a la superficie.}

\item{g}{es el angulo cenital solar.}
}
\value{
Radiacion solar en kwh/m^2
}
\description{
Modelo para obtener la radiacion solar.
}
\examples{
Modelo(1,1,1,1,1,1,1)
}
