% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paqueteRadiacion.R
\docType{package}
\name{paqueteRadiacion}
\alias{paqueteRadiacion}
\alias{paqueteRadiacion-package}
\title{Paquete Radiación
Paquete con el cual se busca evaluar la radiacion solar obtenida por medio de un modelo matematico para ser comparado con los valores originales.}
\description{
Paquete Radiación
Paquete con el cual se busca evaluar la radiacion solar obtenida por medio de un modelo matematico para ser comparado con los valores originales.
}
\author{
Karen Juliana Celis Buitrago, David Steven López Tovar y Andres Camilo Muñoz Usamá
}
